Move GeneratePrompt to utils
Reason: It's functionality we don't want to test as part of models
